pyproject.toml
[project]
name = "api-cash-wise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "django-cors-headers>=4.6.0",
    "Django==5.1.3",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "PyJWT==2.9.0",
    "sqlparse==0.5.1",
    "tzdata==2024.2",
]
